Added DataTableExtensions and SelectTopNRows method there.
Not sure this is the most efficient way to return a new DataTable with the first X rows from another, but it probably helps that it's just a TOP rather than a WHERE-ing. That is, it's not too inefficient to grab the top rows, but I wouldn't want it to get much more complicated than this. Seems most use cases would be better using a DataView or just a subset of Rows from the original table, since grabbing the TOP N and putting them into another table is almost always going to be overkill from an efficiency standpoint. /shrug